+++ packages/atoms/icon/bin/start.sh
@@ -1,17 +1,27 @@
 #!/bin/bash
 
 # variables
+package=$(pwd)
 view=${1:-icon}
 
 for arg in "$@"; do
-  if [[ $arg == --target=* ]]; then 
+  if [[ $arg == --view=* ]]; then 
     view="${arg#*=}"
   fi 
 done
 
 # cleanup
+has_cleaned=false
 cleanup() {
+  if [[ $has_cleaned == true ]]; then 
+    return 
+  fi 
+
+  has_cleaned=true
+  
   echo "[start] cleanup"
+  cd $package
+  
   kill $watch_pid 
   kill $server_pid
 
@@ -29,7 +39,10 @@ fi
 npm run watch &
 watch_pid=$!
 
-npx @papit/server --verbose --location=$(pwd)/views --live $@ &
+# Try running the server with npm workspace
+cd $(npm prefix)
+
+npm run server --include-workspace-root=true --workspace="$(npm prefix)" -- --verbose --location=$package/views/$view --live --nosig "$@" &
 server_pid=$!
 
 # final
